Update to VICE 3.1 and fix AutostartPrgDisk.d64 output path

+++ C64.emu/src/vice/alarm.h
@@ -113,7 +113,7 @@ inline static CLOCK alarm_context_next_pending_clk(alarm_context_t *context)
 inline static void alarm_context_update_next_pending(alarm_context_t *context)
 {
     CLOCK next_pending_alarm_clk = (CLOCK)~0L;
-    unsigned int next_pending_alarm_idx;
+    int next_pending_alarm_idx;
     unsigned int i;
 
     next_pending_alarm_idx = context->next_pending_alarm_idx;
@@ -123,7 +123,7 @@ inline static void alarm_context_update_next_pending(alarm_context_t *context)
 
         if (pending_clk <= next_pending_alarm_clk) {
             next_pending_alarm_clk = pending_clk;
-            next_pending_alarm_idx = i;
+            next_pending_alarm_idx = (int)i;
         }
     }
 
@@ -135,7 +135,7 @@ inline static void alarm_context_dispatch(alarm_context_t *context,
                                           CLOCK cpu_clk)
 {
     CLOCK offset;
-    unsigned int idx;
+    int idx;
     alarm_t *alarm;
 
     offset = (CLOCK)(cpu_clk - context->next_pending_alarm_clk);
@@ -155,11 +155,11 @@ inline static void alarm_set(alarm_t *alarm, CLOCK cpu_clk)
     idx = alarm->pending_idx;
 
     if (idx < 0) {
-        unsigned int new_idx;
+        int new_idx;
 
         /* Not pending yet: add.  */
 
-        new_idx = context->num_pending_alarms;
+        new_idx = (int)(context->num_pending_alarms);
         if (unlikely(new_idx >= ALARM_CONTEXT_MAX_PENDING_ALARMS)) {
             alarm_log_too_many_alarms();
             return;
