+++ C64.emu/src/vice/drive/iec/glue1571.c
@@ -37,5 +37,5 @@ void glue1571_side_set(unsigned int side, struct drive_s *drive)
 
     drive_gcr_data_writeback(drive);
 
-    drive_set_half_track(drive->current_half_track, side, drive);
+    drive_set_half_track(drive->current_half_track, (int)side, drive);
 }
